# docs/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Little Bear RAG Backend API
  version: 0.1.0
  description: >
    P0 OpenAPI contract for Little Bear RAG backend. The contract covers setup,
    authentication, organization, permission, knowledge base, document import,
    indexing operations, query, audit, healthcheck, config, and internal model
    gateway APIs. P0 uses external embedding, rerank, and LLM provider
    endpoints instead of built-in model simulators, and does not include complete
    config approval or rollback workflows.

    Backend contract paths use /internal/v1. OpenAPI, backend routes, frontend
    SDKs, and contract tests should all use this path prefix consistently.
servers:
  - url: http://localhost:8000
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /internal/v1/knowledge-bases:
    get:
      tags: [KnowledgeBase]
      summary: List accessible knowledge bases
      operationId: listKnowledgeBases
      x-token-type: access
      x-required-scope: knowledge_base:read
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/Keyword"
        - $ref: "#/components/parameters/Status"
      responses:
        "200":
          description: Knowledge base list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeBaseListResponse"
        default:
          $ref: "#/components/responses/Error"
...
  /internal/v1/knowledge-bases/{kb_id}/documents:
    get:
      tags: [KnowledgeBase]
      summary: List accessible documents
      operationId: listDocuments
      x-token-type: access
      x-required-scope: document:read
      parameters:
        - $ref: "#/components/parameters/KbId"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/Keyword"
        - $ref: "#/components/parameters/Status"
      responses:
        "200":
          description: Document list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentListResponse"
        default:
          $ref: "#/components/responses/Error"
    post:
      tags: [Import]
      summary: Upload files and create import job
      operationId: createUploadDocumentImport
      x-token-type: access
      x-required-scope: document:import
      x-audit-required: true
      parameters:
        - $ref: "#/components/parameters/KbId"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/DocumentUploadRequest"
      responses:
        "202":
          description: Import job queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportJobResponse"
        default:
          $ref: "#/components/responses/Error"
...
  /internal/v1/documents/{doc_id}:
    get:
      tags: [KnowledgeBase]
      summary: Get accessible document
      operationId: getDocument
      x-token-type: access
      x-required-scope: document:read
      parameters:
        - $ref: "#/components/parameters/DocId"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentResponse"
        default:
          $ref: "#/components/responses/Error"
  /internal/v1/documents/{doc_id}/versions:
    get:
      tags: [KnowledgeBase]
      summary: List accessible document versions
      operationId: listDocumentVersions
      x-token-type: access
      x-required-scope: document:read
      parameters:
        - $ref: "#/components/parameters/DocId"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Document versions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentVersionListResponse"
        default:
          $ref: "#/components/responses/Error"
  /internal/v1/documents/{doc_id}/chunks:
    get:
      tags: [KnowledgeBase]
      summary: List accessible document chunks
      operationId: listDocumentChunks
      x-token-type: access
      x-required-scope: document:read
      x-audit-required: true
      parameters:
        - $ref: "#/components/parameters/DocId"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/Keyword"
        - $ref: "#/components/parameters/Status"
      responses:
        "200":
          description: Chunk list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChunkListResponse"
        default:
          $ref: "#/components/responses/Error"
  /internal/v1/documents/{doc_id}/preview:
    get:
      tags: [KnowledgeBase]
      summary: Preview accessible document
      operationId: getDocumentPreview
      x-token-type: access
      x-required-scope: document:read
      x-audit-required: true
      parameters:
        - $ref: "#/components/parameters/DocId"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/Keyword"
        - $ref: "#/components/parameters/Status"
      responses:
        "200":
          description: Document preview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentPreviewResponse"
        default:
          $ref: "#/components/responses/Error"
  /internal/v1/documents/{doc_id}/sources/{source_id}:
    get:
      tags: [KnowledgeBase]
      summary: Get verifiable citation source
      operationId: getDocumentSource
      x-token-type: access
      x-required-scope: document:read
      x-audit-required: true
      parameters:
        - $ref: "#/components/parameters/DocId"
        - $ref: "#/components/parameters/SourceId"
      responses:
        "200":
          description: Citation source
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CitationSourceResponse"
        default:
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 20
    Keyword:
      name: keyword
      in: query
      schema:
        type: string
    Status:
      name: status
      in: query
      schema:
        type: string
    KbId:
      name: kb_id
      in: path
      required: true
      schema:
        type: string
    DocId:
      name: doc_id
      in: path
      required: true
      schema:
        type: string
    SourceId:
      name: source_id
      in: path
      required: true
      schema:
        type: string
...
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    RequestId:
      type: string
      examples: ["req_001"]
...
    AnyObject:
      type: object
      additionalProperties: true
    Pagination:
      type: object
      required: [page, page_size, total]
      properties:
        page:
          type: integer
        page_size:
          type: integer
        total:
          type: integer
    ErrorResponse:
      type: object
      required: [request_id, error_code, message, retryable]
      properties:
        request_id:
          $ref: "#/components/schemas/RequestId"
        error_code:
          type: string
        message:
          type: string
        stage:
          type: string
        retryable:
          type: boolean
        details:
          $ref: "#/components/schemas/AnyObject"
...
    Visibility:
      type: string
      enum: [department, enterprise]
...
    ImportJobStatus:
      type: string
      enum: [queued, running, retrying, partial_success, success, failed, cancelled]
...
    Citation:
      type: object
      required: [source_id, doc_id, document_version_id, title, page_start, page_end]
      properties:
        source_id:
          type: string
        doc_id:
          type: string
        document_version_id:
          type: string
        title:
          type: string
        page_start:
          type: integer
        page_end:
          type: integer
        score:
          type: number
...
    KnowledgeBaseListItem:
      type: object
      required:
        - id
        - name
        - status
        - owner_department_id
        - kb_visibility
        - default_document_visibility
        - default_document_owner_department_id
      properties:
        id:
          type: string
        name:
          type: string
        status:
          type: string
          enum: [active, disabled, archived]
        owner_department_id:
          type: string
        owner_department_name:
          type: [string, "null"]
        kb_visibility:
          type: string
          enum: [enterprise, department_acl, private]
        default_document_visibility:
          type: string
          enum: [department, enterprise]
        default_document_owner_department_id:
          type: string
        default_document_owner_department_name:
          type: [string, "null"]
...
    KnowledgeBaseListResponse:
      type: object
      required: [request_id, data, pagination]
      properties:
        request_id:
          $ref: "#/components/schemas/RequestId"
        data:
          type: array
          items:
            $ref: "#/components/schemas/KnowledgeBaseListItem"
        pagination:
          $ref: "#/components/schemas/Pagination"
...
    Document:
      type: object
      required: [id, kb_id, title, lifecycle_status, index_status, owner_department_id, visibility]
      properties:
        id:
          type: string
        kb_id:
          type: string
        folder_id:
          type: [string, "null"]
        title:
          type: string
        lifecycle_status:
          type: string
          enum: [draft, active, archived, deleted]
        index_status:
          type: string
          enum: [none, indexing, indexed, index_failed, blocked]
        owner_department_id:
          type: string
        visibility:
          $ref: "#/components/schemas/Visibility"
        current_version_id:
          type: [string, "null"]
...
    DocumentResponse:
      type: object
      properties:
        request_id:
          $ref: "#/components/schemas/RequestId"
        data:
          $ref: "#/components/schemas/Document"
    DocumentListResponse:
      type: object
      properties:
        request_id:
          $ref: "#/components/schemas/RequestId"
        data:
          type: array
          items:
            $ref: "#/components/schemas/Document"
        pagination:
          $ref: "#/components/schemas/Pagination"
    DocumentVersion:
      type: object
      properties:
        id:
          type: string
        document_id:
          type: string
        version_no:
          type: integer
        status:
          type: string
    DocumentVersionListResponse:
      type: object
      required: [request_id, data, pagination]
      properties:
        request_id:
          $ref: "#/components/schemas/RequestId"
        data:
          type: array
          items:
            $ref: "#/components/schemas/DocumentVersion"
        pagination:
          $ref: "#/components/schemas/Pagination"
    Chunk:
      type: object
      properties:
        id:
          type: string
        document_id:
          type: string
        document_version_id:
          type: string
        text_preview:
          type: string
        page_start:
          type: integer
        page_end:
          type: integer
        status:
          type: string
        ordinal:
          type: integer
    ChunkListResponse:
      type: object
      required: [request_id, data, pagination]
      properties:
        request_id:
          $ref: "#/components/schemas/RequestId"
        data:
          type: array
          items:
            $ref: "#/components/schemas/Chunk"
        pagination:
          $ref: "#/components/schemas/Pagination"
...
    CitationSource:
      type: object
      required:
        - source_id
        - doc_id
        - document_version_id
        - title
        - text
        - text_preview
        - ordinal
        - text_status
      properties:
        source_id:
          type: string
        doc_id:
          type: string
        document_version_id:
          type: string
        title:
          type: string
        text:
          type: string
        text_preview:
          type: string
        page_start:
          type: [integer, "null"]
        page_end:
          type: [integer, "null"]
        ordinal:
          type: integer
        heading_path:
          type: [string, "null"]
        source_offsets:
          type: [object, "null"]
          additionalProperties: true
        text_status:
          type: string
          enum: [object, preview_only, object_unavailable]
    CitationSourceResponse:
      type: object
      properties:
        request_id:
          $ref: "#/components/schemas/RequestId"
        data:
          $ref: "#/components/schemas/CitationSource"
    DocumentPreviewResponse:
      type: object
      properties:
        request_id:
          $ref: "#/components/schemas/RequestId"
        data:
          type: object
          properties:
            doc_id:
              type: string
            title:
              type: string
            preview:
              type: string
            citations:
              type: array
              items:
                $ref: "#/components/schemas/Citation"
...
    DocumentUploadRequest:
      type: object
      required: [files]
      properties:
        files:
          type: array
          items:
            type: string
            format: binary
        owner_department_id:
          type: string
        visibility:
          $ref: "#/components/schemas/Visibility"
        folder_id:
          type: [string, "null"]
        idempotency_key:
          type: string
...
    ImportJob:
      type: object
      required: [id, kb_id, status, stage]
      properties:
        id:
          type: string
        kb_id:
          type: [string, "null"]
        job_type:
          type: [string, "null"]
        status:
          $ref: "#/components/schemas/ImportJobStatus"
        stage:
          type: string
          enum: [validate, parse, clean, chunk, embed, index, publish, cleanup, finished]
        document_ids:
          type: array
          items:
            type: string
        error_summary:
          type: [string, "null"]
...
    ImportJobResponse:
      type: object
      required: [request_id, data]
      properties:
        request_id:
          $ref: "#/components/schemas/RequestId"
        data:
          $ref: "#/components/schemas/ImportJob"
...